# docs/chatgpt/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Statement2Muster DATEV & BMD Accounting Action",
    "description": "Official accounting conversion action for Statement2Muster. Enables ChatGPT to turn bank statements (PDF, CSV, image or text from American Express, Wise, PayPal, Sparkasse, Deutsche Bank, Erste Bank, Raiffeisen, etc.) into certified DATEV EXTF (Buchungsstapel Version 700 Windows-1252) and Austrian BMD NTCS 5.1 CSV files. Features automated debit/credit determination (Soll/Haben), German bank account mapping (SKR03/SKR04/BMD), financial reconciliation summary, and 100% DSGVO & Zero Durable Retention (data processed strictly in volatile RAM with 30-minute auto-purge).\n",
    "version": "2.1.0"
  },
  "servers": [
    {
      "url": "https://api.statement2muster.com",
      "description": "Production API Server (Hetzner ISO 27001 Certified)"
    }
  ],
  "paths": {
    "/v1/gpt/convert": {
      "post": {
        "operationId": "convertStatement",
        "summary": "Convert bank statement transactions to certified DATEV EXTF or BMD NTCS format",
        "description": "Converts extracted transactions, raw statement text/CSV, or base64 PDF into official German DATEV EXTF 700 (Windows-1252, SKR03/SKR04) or Austrian BMD NTCS 5.1 format. Generates financial reconciliation summary (Debit/Credit/Balance), base64 file payload for immediate inline chat access, and an ephemeral download link. Free tier includes 3 conversions per session; unlimited with Starter (€4.90) or Pro (€29) license.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GptConvertRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful conversion with download URL, base64 payload, and accounting turnover summary",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GptConvertResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters or empty data"
          },
          "413": {
            "description": "Payload or file size exceeds allowed limits"
          },
          "422": {
            "description": "Unprocessable transaction data or zero bookings extracted"
          }
        }
      }
    },
    "/v1/gpt/check-license": {
      "post": {
        "operationId": "checkLicense",
        "summary": "Check license key or user email subscription status",
        "description": "Verifies whether a user has an active Statement2Muster Starter, Pro, or Lifetime license using their license key, Stripe checkout session ID, or registered email.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckLicenseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "License validation result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckLicenseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/gpt/download/{download_id}": {
      "get": {
        "operationId": "downloadConvertedFile",
        "summary": "Download generated accounting CSV file from RAM cache",
        "description": "Streams the generated Windows-1252 encoded DATEV or BMD file directly from volatile memory. Available for 30 minutes following conversion.\n",
        "parameters": [
          {
            "name": "download_id",
            "in": "path",
            "required": true,
            "description": "Unique ephemeral download token returned by convertStatement",
            "schema": {
              "type": "string",
              "example": "s2m_gpt_7c3a1b2c3d4e5f6"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Windows-1252 encoded CSV file with attachment disposition",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "Download link expired or purged from RAM per Zero Durable Retention policy"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "GptTransactionItem": {
        "type": "object",
        "required": [
          "booking_date",
          "amount",
          "description"
        ],
        "properties": {
          "booking_date": {
            "type": "string",
            "description": "Booking date in ISO YYYY-MM-DD or European DD.MM.YYYY format",
            "example": "2026-03-15"
          },
          "value_date": {
            "type": "string",
            "description": "Optional valuta/value date (YYYY-MM-DD or DD.MM.YYYY)",
            "example": "2026-03-16"
          },
          "amount": {
            "type": "number",
            "format": "float",
            "description": "Signed transaction amount. Negative value represents outgoing expense / debit (Lastschrift / Ausgabe). Positive value represents incoming credit / revenue (Gutschrift / Einnahme).\n",
            "example": -145.2
          },
          "currency": {
            "type": "string",
            "description": "ISO 4217 Currency Code (EUR, USD, CHF, GBP)",
            "default": "EUR",
            "example": "EUR"
          },
          "description": {
            "type": "string",
            "description": "Payment purpose, beneficiary or booking text (DATEV Buchungstext)",
            "example": "AWS EMEA Cloud Services"
          },
          "reference": {
            "type": "string",
            "description": "Invoice number or transaction reference (DATEV Belegfeld 1)",
            "example": "INV-99102-DE"
          },
          "contra_account": {
            "type": "string",
            "description": "Optional pre-assigned contra account / Gegenkonto (e.g. 4900 for SKR03, 6800 for SKR04)",
            "example": "4900"
          }
        }
      },
      "GptConvertRequest": {
        "type": "object",
        "properties": {
          "session_id": {
            "type": "string",
            "description": "Unique ChatGPT session or conversation identifier to track free tier usage",
            "example": "chatgpt-sess-88123"
          },
          "email": {
            "type": "string",
            "description": "Optional user email address to verify active Pro subscription or track quota",
            "example": "founder@startup.de"
          },
          "license_key": {
            "type": "string",
            "description": "Optional Statement2Muster Pro / Starter / Lifetime license key or Stripe checkout session ID",
            "example": "cs_live_12345abcdef"
          },
          "bank_name": {
            "type": "string",
            "description": "Name of card issuer or bank (e.g. American Express, Wise, PayPal, Sparkasse, Volksbank)",
            "default": "Bank Statement",
            "example": "American Express Business"
          },
          "export_format": {
            "type": "string",
            "enum": [
              "datev",
              "bmd",
              "muster_csv"
            ],
            "default": "datev",
            "description": "Target accounting format (datev = DATEV EXTF 700 Windows-1252, bmd = BMD NTCS 5.1)"
          },
          "default_bank_account": {
            "type": "string",
            "description": "Bank ledger account number. Default for DATEV is 1200 (SKR03) or 1800 (SKR04). Default for BMD is 2800.",
            "example": "1200"
          },
          "transactions": {
            "type": "array",
            "description": "Array of extracted transactions from document",
            "items": {
              "$ref": "#/components/schemas/GptTransactionItem"
            }
          },
          "raw_content": {
            "type": "string",
            "description": "Optional raw CSV or plaintext text if client did not extract individual items"
          },
          "file_base64": {
            "type": "string",
            "description": "Optional base64-encoded PDF or CSV document for server-side supervised extraction"
          },
          "filename": {
            "type": "string",
            "description": "Original statement filename",
            "example": "amex_statement_march_2026.pdf"
          }
        }
      },
      "FinancialSummary": {
        "type": "object",
        "properties": {
          "transaction_count": {
            "type": "integer",
            "example": 42
          },
          "total_debit": {
            "type": "number",
            "format": "float",
            "description": "Total sum of expenses / outgoing amounts (negative)",
            "example": -4890.5
          },
          "total_credit": {
            "type": "number",
            "format": "float",
            "description": "Total sum of revenues / incoming amounts (positive)",
            "example": 12500.0
          },
          "net_balance": {
            "type": "number",
            "format": "float",
            "description": "Net period turnover (Credit + Debit)",
            "example": 7609.5
          },
          "currency": {
            "type": "string",
            "example": "EUR"
          },
          "date_from": {
            "type": "string",
            "example": "2026-03-01"
          },
          "date_to": {
            "type": "string",
            "example": "2026-03-31"
          }
        }
      },
      "FreeTierStatus": {
        "type": "object",
        "properties": {
          "is_unlimited": {
            "type": "boolean",
            "description": "True if covered by active paid subscription",
            "example": false
          },
          "plan": {
            "type": "string",
            "description": "Active plan code (starter, pro, lifetime) if licensed",
            "example": "pro"
          },
          "conversions_used": {
            "type": "integer",
            "description": "Number of free conversions already used",
            "example": 1
          },
          "remaining_free_conversions": {
            "type": "integer",
            "description": "Number of free conversions remaining",
            "example": 2
          },
          "max_free_conversions": {
            "type": "integer",
            "description": "Maximum allowed free conversions without license",
            "example": 3
          },
          "limit_reached": {
            "type": "boolean",
            "description": "True if free tier limit has been exhausted",
            "example": false
          }
        }
      },
      "GptConvertResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "success",
              "warning",
              "error",
              "limit_reached"
            ],
            "example": "success"
          },
          "message": {
            "type": "string",
            "example": "Erfolgreich 42 Buchungssätze in DATEV konvertiert."
          },
          "export_format": {
            "type": "string",
            "example": "datev"
          },
          "filename": {
            "type": "string",
            "example": "DATEV_EXTF_American_Express_20260326_120000.csv"
          },
          "download_id": {
            "type": "string",
            "example": "s2m_gpt_7c3a1b2c3d4e5f6"
          },
          "download_url": {
            "type": "string",
            "format": "uri",
            "description": "Clickable temporary download link (valid for 30 minutes in RAM)",
            "example": "https://api.statement2muster.com/v1/gpt/download/s2m_gpt_7c3a1b..."
          },
          "file_base64": {
            "type": "string",
            "description": "Base64-encoded CSV content for inline presentation or Code Interpreter in ChatGPT"
          },
          "expires_in_minutes": {
            "type": "integer",
            "example": 30
          },
          "summary": {
            "$ref": "#/components/schemas/FinancialSummary"
          },
          "preview_csv": {
            "type": "string",
            "description": "First 5 lines of the generated accounting file for verification"
          },
          "free_tier_status": {
            "$ref": "#/components/schemas/FreeTierStatus"
          },
          "notes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "upgrade_info": {
            "type": "object",
            "description": "Stripe checkout links to purchase or upgrade"
          }
        }
      },
      "CheckLicenseRequest": {
        "type": "object",
        "required": [
          "license_key"
        ],
        "properties": {
          "license_key": {
            "type": "string",
            "description": "License key, Stripe checkout session ID, or user email",
            "example": "cs_live_12345abcdef"
          },
          "email": {
            "type": "string",
            "description": "Optional user email",
            "example": "user@example.com"
          }
        }
      },
      "CheckLicenseResponse": {
        "type": "object",
        "required": [
          "valid",
          "message"
        ],
        "properties": {
          "valid": {
            "type": "boolean",
            "example": true
          },
          "plan": {
            "type": "string",
            "example": "pro"
          },
          "message": {
            "type": "string",
            "example": "Valid license confirmed (PRO)."
          },
          "upgrade_info": {
            "type": "object"
          }
        }
      }
    }
  }
}